NPCバトルステージを別ファイルに切り出した
